# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/api/v1/timezone": {
            "get": {
                "description": "Returns the current configured timezone",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "timezone"
                ],
                "summary": "Get current timezone",
                "responses": {
                    "200": {
                        "description": "{ success: true, data: TimezoneResponse }",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "{ success: false, error: string }",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "post": {
                "description": "Updates the configured timezone",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "timezone"
                ],
                "summary": "Update timezone",
                "parameters": [
                    {
                        "description": "Timezone data",
                        "name": "timezone",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.TimezoneUpdateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{ success: true, data: TimezoneResponse, message: string }",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "{ success: false, error: string }",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "{ success: false, error: string }",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "api.TimezoneUpdateRequest": {
            "description": "Request payload for updating timezone.",
            "type": "object",
            "properties": {
                "location": {
                    "type": "string"
                }
            }
        },
...
    }
}